Optimized Performance
The get(Sha256Hash hash) functions try to get storedBlock from the cache first. If storedBlock is found the storedBlock is returned from the cache, otherwise if not find then it computes the storedBlock from byteArrayToStoredblock. ByteArrayToStoredBlock processing time is longer.  The current function processes the BridgeSync Transaction much faster.