Set upper limit in voteFeeperKbChange
Defined a new BridgeConstant with the upper limit for the feePerKb for each network. Using said value in voteFeePerKbChange method, rejecting any value above it.

Confirmed it's not a hard fork using Blockchain verificator tool to check that voteFeePerKbChange method was never called in mainnet or testnet with a value above the maximum defined