Reorder the executedTransactions and transactionReceipts in BlockResult
In case of different transactions are executed in parallel, the 'executedTransactions' and 'transactionReceipts' lists in the BlockResult, being populated on the fly, receive execution results in an undeterminist order.
So, after all execution have completed, they are reordered to match the "submission order" (as ordered in the block to be validated).