Fix BridgeState usage after Papyrus activation
Bridge state is a DTO used to access vital information of the Bridge storage. This includes transactions to be releases, transactions waiting for confirmations, or signatures, among other things.
The Bridge storage was modified during the changes for Papyrus, without contemplating the usage from BridgeState.

This commit adds the activations object to BridgeState to let it determine which serialization method to use, depending on said activations.

Adding to that, this commit also fixes an issue with a test where the Bridge Balance was not set, therefore it assumed all the RBTC were distributed and could not allocate more.