Added a new way to transits to StartRoundSync from genesis, also the maxAmountOfHeaders is stored as a instance variable so when the peer receives a list of block headers in CommonAncestorSearch state, it checks against maxAmountOfHeaders if its correct (#1271)
Added a new way to transits to StartRoundSync from genesis. If the best block headers of the light peers is Genesis, its transits directly to StartRoundSync otherwise it transits to SearchCommonAncestor. The last way is when the peers common ancestor is Genesis.
The maxAmountOfHeaders is stored as a instance variable so when the peer receives a list of block headers in CommonAncestorSearch state, it checks against maxAmountOfHeaders if its correct