Modify the value saved in BridgeStorageProvider so that the second value is a simple value.
Change the information stored in BridgeStorageProvider. Now go on to save the hash of the derivation arguments with a 1-byte dummy value. This value is only used to save content together with the hash. It is only interesting to know if the hash of derivation arguments that is saved as key was calculated, it really does not matter what content is saved with this value.