Update lock_btc event
After RSKIP 170 activation sender address is no longer registered and peg-in protocol version is added. Also btc tx hash is now a topic.

Removed logic to put "Undetermined" in sender field when the sender could not be obtained, since after HF activation the new event will be used that does not have this field.