Depth limit for Receive Header.
Check the depth of the block, taking into account the depth in the storage of the previous block.
If the depth of the block is too old, then it returns error. If it can't find the previous block in storage, then it also returns error.
The new errors that are introduced here are:

-2 -> RECEIVE_HEADER_BLOCK_TOO_OLD
-3 -> RECEIVE_HEADER_CANT_FOUND_PREVIUOS_BLOCK