Add new check for peg-out minimum considering fees
A peg-out transaction has a minimum established in the Bridge constants.
This value doesn't take into account the changing nature of the fees.
Ths commit adds a validation to calculate the approximated cost of a peg-out with the current fees per KB of the network.
The cost is calculated considering a transaction with two inputs and two outputs, which is the most common situation ATM.

A peg-out transaction should also generate value for the user, this is, if a peg-out tx costs 400k satoshis and the user is attempting to peg-out 450k, the remainder (50k) won't be of much use for the user. In order to avoid this situation the Bridge now includes a percentage above the fee cost that the peg-out should include, if this is amount is not attained, the peg-out is rejected and funds returned.