Rskip351+rskip144+precompiled contracts (#1994)
From now on, whenever a tx execution touches a Precompiled contract, it goes to the sequential sublist. To do that, I added a new flag called PrecompiledHasBeenCalled in both the Program and the TransactionExecutor classes. Furthermore, I tested the flag in each class and added some DSL tests for a more complete testing.