Adding skeleton for integration tests for SnapshotSynch
In order to be able to have a good test for snapshot sync, we need
to have way to spin up two nodes in a configurable manner so we can
validate if these nodes are able to sync between them.

This first commit achieves this creating the classes for it. We still
need to add the real logic and the configuration for server and client, this
branch still a WIP and it's a draft.