Refactored the NodeIntegrationTestCommandLine
In order to validate properly the snap synch, we needed a process that
executed in background and we could get the output generated after it
had finished. This process also needed to be killed only after some time
to be decided by the test. Due that, we had to refactor the code to allow this.

Now we are ready to configure the node with different conf file and make them sync.