[WIP] Refactored the command line in order to re-use it
So we can use the different command lines, it was created a command line class base
and different classes according the template design pattern. Maybe we still need to
do some refactor to create a factory, but the way it is now, it's decent and easy to
add a new command.

Now we will import the exported blocks and create the integrated test with server and
client node.