Remove dependency between BridgeConstants and Federation classes
- Adding changes to the tests to adapt it to new way to get Genesis Federation.
- Creating a new constant called GENESIS_FEDERATION_CREATION_BLOCK_NUMBER in FederationTestUtils.java and this same variable/constant in FederationSupport#getGenesisFederation
- Improving the readability in the statement lambda in BridgeUtils#isFromGenesisFederation
- New test to make sure BrigdeConstants has the correct GenesisFederationCreationTime according to the network.
- Improving Utils classes:
FederationTestUtils:
adding final keyword at level class to avoid this class could be inherited and to improve the class performance.
adding a private constructor to avoid this utility class could be instantiated

BridgeUtils: adding final keyword at level class to avoid this class could be inherited and to improve the class performance.