Unit Tests for WhitelistConstants
Following the Bridge refactors, we already migrated the WhitelistConstants to their classes to break the high coupling Whitelist had with the Bridge objects, however, it’s missing add coverage, so we need to create Unit Tests to get reliability your new code doesn't break existing functionality when a change is made to the application.

Create Test Cases for all Networks such as MainNet, TestNet, and RegTest.

Addressing Review Comments

- Improving WhitelistConstants: test and production code for all Networks.