Whitelist Support Test
Add Unit Tests for WhitelistSupportImpl making sure we are simulating real scenarios.

Renaming Test Methods

- Renaming Test Methods to get closer what they do.

Add UTs for verifyLockSenderIsWhitelisted in WhitelistSupportImplTest

As part of Bridge refactors, this is time to work on Whitelist process, so we need to extract this from the Bridge and create its own objects to decouple the process.

Create UTs for verifyLockSenderIsWhitelisted to increase coverage on WhitelistSupportImplTest.

Verify address as consumed after verification