LockingCapConstants Test
Following the Bridge refactors, we already migrated the LockingCapConstants to their classes to break the high coupling LockingCap had with the Bridge objects, however, it’s missing add coverage, so we need to create Unit Tests to get reliability that your new code doesn't break existing functionality when a change is made to the application.

Create Unit Tests for all methods in LockingCapConstants
Create Test Cases for all Networks such as MainNet, TestNet, and RegTest.