Fix PowpegMigrationTest after LockingCap refactor
Following the Bridge refactors, we already migrated the LockingCap processes to their classes to break the high coupling LockingCap had with the Bridge objects. After refactoring, showed up different issues in the tests and it is time to fix them.

Fix the PowpegMigrationTest class.
Get rid of setting the LockingCap since it is not mandatory in the test. From now on, the test will use the LockingCap default value, i.e., the LockingCapConstants.initialValue().