Initcode validation for opcodes contract creation
- In order to have the exact same logic for the initcode cost calculation,
we introduced some new classes that can be easily reused by the TransactionExecutor and VM classes. This way, if this logic change,
we won't have to change the code in multiple places.

It was also introduced an interface for it, so we can expand this idea and
in the future, be able to calculate total transactions just calling a method
to calculate it independent of the logic used.

- Adding scenarios of test for the contract creation with DSL
- The validation is based on the RSKIP that enables it