Feature/initcode size limit using transaction creations (#2665)
* Initial constants and config for the max initcode size

In order to implement the verification of initcode size, necessary for the RSKIP438,
we need the constant value with the max allowed value. So we can start to validate
and use it on the logic of the verification

- Adding scenarios of test for the contract creation with DSL
- The validation for the initCode size test was added
- The validation is based on the RSKIP that enables it
- Validation added for the contractCreation during transaction execution

Finished the implementation of transactionCost and initcode size validation

- All the logic related with transaction cost was changed to meet the new cost criteria
- More tests were added to be sure that the flow is following the RSKIP438
- Tests were added to validate that this cost and validation doesn't impact before reach the activation height

Applied some suggestions from review

Adding tests to validate the failure if we don't have enough gas to cover initCode

* Refactor from the Initcode cost calculation

In order to have the exact same logic for the initcode cost calculation,
we introduced some new classes that can be easily reused by the TransactionExecutor and VM classes. This way, if this logic change, we won't have to change the code in multiple places.

It was also introduced an interface for it, so we can expand this idea and
in the future, be able to calculate total transactions just calling a method
to calculate it independent of the logic used.

* Applying some more suggestions from review

Co-Authored-By: Nazaret García Revetria <nazaret@iovlabs.org>
Co-Authored-By: Vovchyk <volodymyr.kravets@outlook.com>