Configure spotless to conform with the project's Checkstyle rules
The previous configuration of spotless wasn't compatible with the current rules for checkstyle. As a consequence, the code formatted by spotless would not pass the linting rules of checkstyle.

This commit changes the configuration to use the eclipse formatter, which allows for custom configuration. That way, we can set up both tools so that they agree with eachother.

See https://github.com/eclipse-jdt/eclipse.jdt.core/blob/master/org.eclipse.jdt.core/formatter/org/eclipse/jdt/core/formatter/DefaultCodeFormatterConstants.java for a complete list of supported configs (it's a lot!)