Instrument DifficultyCalculator.java to dump relevant metrics to a CSV
This will be used to evaluate the performance of the new difficulty calculation algorithm proposed on RSKIP517. For every new block, when the block difficulty is computed, the metrics for current and average block difficulty are dumped to a CSV file. This can be used to compare the current algorithm and the one proposed on RSKIP517 side by side.